fixing some generics issues and class cast exceptions due to use of Value instead of raw wrapped value - all tests in NestedFilterTest now pass
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/branches/bigbang@656041 13f79535-47bb-0310-9956-ffa450edef68

+++ xdbm-search/src/main/java/org/apache/directory/server/xdbm/search/impl/OrEvaluator.java
@@ -73,8 +73,8 @@ public class OrEvaluator implements Evaluator<OrNode, ServerEntry>
         {
             public int compare( Evaluator<? extends ExprNode, ServerEntry> e1, Evaluator<? extends ExprNode, ServerEntry> e2 )
             {
-                int scanCount1 = ( Integer ) e1.getExpression().get( "count" );
-                int scanCount2 = ( Integer ) e2.getExpression().get( "count" );
+                long scanCount1 = ( Long ) e1.getExpression().get( "count" );
+                long scanCount2 = ( Long ) e2.getExpression().get( "count" );
 
                 if ( scanCount1 == scanCount2 )
                 {
